pyproject.toml
[project]
name = "ecosystem-alfa-tc"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-stubs>=3.16.8",
    "drf-spectacular>=0.29.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "python-dotenv>=1.2.1",
]
